fix default configuration extraction
