+++ custom_components/aqua_temp/manifest.json
@@ -8,5 +8,5 @@
   "iot_class": "cloud_polling",
   "issue_tracker": "https://github.com/radical-squared/aquatemp/issues",
   "requirements": [],
-  "version": "3.0.11"
+  "version": "3.0.11b"
 }
